# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPU Fleet Commander API",
    "description": "Control Plane and Telemetry Ingestion Hub for distributed worker nodes.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/nodes": {
      "post": {
        "summary": "Register a worker node",
        "operationId": "register_node_v1_nodes_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all registered worker nodes",
        "operationId": "list_nodes_v1_nodes_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeResponse"
                  },
                  "title": "Response List Nodes V1 Nodes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nodes/{node_id}/heartbeat": {
      "post": {
        "summary": "Ingest heartbeat signal",
        "operationId": "process_heartbeat_v1_nodes__node_id__heartbeat_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks": {
      "post": {
        "summary": "Create a computational task",
        "operationId": "create_task_v1_tasks_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/dispatch": {
      "post": {
        "summary": "Dispatch task to online node",
        "operationId": "dispatch_task_v1_tasks__task_id__dispatch_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/transition": {
      "post": {
        "summary": "Transition running task status",
        "operationId": "transition_task_v1_tasks__task_id__transition_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "summary": "Retrieve task details by ID",
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nodes/{node_id}/telemetry": {
      "post": {
        "summary": "Ingest node metrics telemetry",
        "operationId": "ingest_telemetry_v1_nodes__node_id__telemetry_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryIngestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve recent telemetry metrics for a worker node",
        "operationId": "get_node_telemetry_v1_nodes__node_id__telemetry_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of metrics to return",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of metrics to return"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetryResponse"
                  },
                  "title": "Response Get Node Telemetry V1 Nodes  Node Id  Telemetry Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/metrics": {
      "get": {
        "summary": "Export fleet & system metrics for Prometheus",
        "operationId": "get_prometheus_metrics_v1_metrics_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Access API Key for control plane authentication",
              "title": "X-Api-Key"
            },
            "description": "Access API Key for control plane authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "API Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Serve Web Dashboard",
        "operationId": "serve_dashboard__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DispatchRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Node Id",
            "description": "Target node ID to dispatch the task to"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "DispatchRequest",
        "description": "Request body for dispatching a task to a specific node."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NodeRegisterRequest": {
        "properties": {
          "hostname": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Hostname",
            "description": "Unique hostname of the node"
          },
          "hardware_specs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hardware Specs",
            "description": "JSON representing hardware details (GPU info, memory, etc.)"
          }
        },
        "type": "object",
        "required": [
          "hostname",
          "hardware_specs"
        ],
        "title": "NodeRegisterRequest"
      },
      "NodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "hardware_specs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hardware Specs"
          },
          "last_heartbeat": {
            "type": "string",
            "format": "date-time",
            "title": "Last Heartbeat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hostname",
          "status",
          "hardware_specs",
          "last_heartbeat"
        ],
        "title": "NodeResponse"
      },
      "NodeStatus": {
        "type": "string",
        "enum": [
          "ONLINE",
          "OFFLINE",
          "MAINTENANCE"
        ],
        "title": "NodeStatus"
      },
      "TaskCreateRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Idempotency Key",
            "description": "Client-provided idempotency token"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "Parameters required for executing the task"
          }
        },
        "type": "object",
        "required": [
          "idempotency_key",
          "payload"
        ],
        "title": "TaskCreateRequest"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "retries": {
            "type": "integer",
            "title": "Retries"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "payload",
          "status",
          "retries",
          "created_at"
        ],
        "title": "TaskResponse"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "TaskStatus"
      },
      "TelemetryIngestRequest": {
        "properties": {
          "cpu_usage": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cpu Usage",
            "description": "CPU usage percentage"
          },
          "gpu_usage": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Gpu Usage",
            "description": "GPU usage percentage"
          },
          "temperature": {
            "type": "number",
            "minimum": -273.15,
            "title": "Temperature",
            "description": "Temperature in Celsius"
          }
        },
        "type": "object",
        "required": [
          "cpu_usage",
          "gpu_usage",
          "temperature"
        ],
        "title": "TelemetryIngestRequest"
      },
      "TelemetryResponse": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "cpu_usage": {
            "type": "number",
            "title": "Cpu Usage"
          },
          "gpu_usage": {
            "type": "number",
            "title": "Gpu Usage"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "timestamp",
          "cpu_usage",
          "gpu_usage",
          "temperature"
        ],
        "title": "TelemetryResponse"
      },
      "TransitionRequest": {
        "properties": {
          "target_status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Target status (COMPLETED or FAILED)"
          }
        },
        "type": "object",
        "required": [
          "target_status"
        ],
        "title": "TransitionRequest",
        "description": "Request body for transitioning a task to a new status."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}